src/fsverity: split out FsVerityHashValue and impls
Split FsVerityHashValue and its two implementations out to a separate
file.  Right now it's not a lot of code, but it's going to get
substantially larger in the next commit.
