+++ src/fsverity/mod.rs
@@ -1,36 +1,12 @@
 mod digest;
+mod hashvalue;
 mod ioctl;
 
 use std::{io::Error, os::fd::AsFd};
 
-use sha2::{digest::FixedOutputReset, digest::Output, Digest, Sha256, Sha512};
 use thiserror::Error;
 
-pub trait FsVerityHashValue
-where
-    Self: Eq + AsRef<[u8]> + Clone,
-    Self: From<Output<Self::Digest>>,
-{
-    type Digest: Digest + FixedOutputReset + std::fmt::Debug;
-    const ALGORITHM: u8;
-    const EMPTY: Self;
-}
-
-pub type Sha256HashValue = [u8; 32];
-
-impl FsVerityHashValue for Sha256HashValue {
-    type Digest = Sha256;
-    const ALGORITHM: u8 = 1;
-    const EMPTY: Self = [0; 32];
-}
-
-pub type Sha512HashValue = [u8; 64];
-
-impl FsVerityHashValue for Sha512HashValue {
-    type Digest = Sha512;
-    const ALGORITHM: u8 = 2;
-    const EMPTY: Self = [0; 64];
-}
+pub use hashvalue::{FsVerityHashValue, Sha256HashValue, Sha512HashValue};
 
 /// Measuring fsverity failed.
 #[derive(Error, Debug)] // can't derive PartialEq because of std::io::Error
